Работа над замечаниями. Изменение структуры модулей проекта. Созданы модули: CORE, UI, UTILS. Остальные будем добавлять по ходу работы.
