+++ build.gradle.kts
@@ -7,6 +7,7 @@ buildscript {
     dependencies {
         classpath("com.android.tools.build:gradle:${LibVersion.GRADLE_VERSION}")
         classpath("org.jetbrains.kotlin:${LibVersion.GRADLE_PLUGIN_VERSION}")
+        classpath("org.jetbrains.kotlin:kotlin-gradle-plugin:1.5.31")
 
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
