+++ core/build.gradle.kts
