+++ settings.gradle
@@ -8,8 +8,6 @@ dependencyResolutionManagement {
 }
 rootProject.name = "Mobile Waiter"
 include ':app'
-include ':domain'
 include ':utils'
-include ':screenbills'
-include ':screentables'
-include ':screenbill'
+include ':core'
+include ':ui'
